Refactor location of status_codes
