+++ src/backend/InvenTree/order/api.py
@@ -30,14 +30,6 @@ from InvenTree.filters import SEARCH_ORDER_FILTER, SEARCH_ORDER_FILTER_ALIAS
 from InvenTree.helpers import DownloadFile, str2bool
 from InvenTree.helpers_model import construct_absolute_url, get_base_url
 from InvenTree.mixins import CreateAPI, ListAPI, ListCreateAPI, RetrieveUpdateDestroyAPI
-from InvenTree.status_codes import (
-    PurchaseOrderStatus,
-    PurchaseOrderStatusGroups,
-    ReturnOrderLineStatus,
-    ReturnOrderStatus,
-    SalesOrderStatus,
-    SalesOrderStatusGroups,
-)
 from order import models, serializers
 from order.admin import (
     PurchaseOrderExtraLineResource,
@@ -47,6 +39,14 @@ from order.admin import (
     SalesOrderExtraLineResource,
     SalesOrderLineItemResource,
     SalesOrderResource,
+)
+from order.status_codes import (
+    PurchaseOrderStatus,
+    PurchaseOrderStatusGroups,
+    ReturnOrderLineStatus,
+    ReturnOrderStatus,
+    SalesOrderStatus,
+    SalesOrderStatusGroups,
 )
 from part.models import Part
 from users.models import Owner
