+++ src/backend/InvenTree/plugin/samples/integration/transition.py
@@ -2,8 +2,8 @@
 
 from common.notifications import trigger_notification
 from generic.states import TransitionMethod
-from InvenTree.status_codes import ReturnOrderStatus
 from order.models import ReturnOrder
+from order.status_codes import ReturnOrderStatus
 from plugin import InvenTreePlugin
